requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.3.1
    # via
    #   tensorboard
    #   tensorflow
affine==2.4.0
    # via rasterio
astunparse==1.6.3
    # via tensorflow
attrs==25.3.0
    # via
    #   fiona
    #   rasterio
cachetools==5.5.2
    # via google-auth
certifi==2025.10.5
    # via
    #   fiona
    #   pyogrio
    #   pyproj
    #   rasterio
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.1.8
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   rasterio
click-plugins==1.1.1.2
    # via
    #   fiona
    #   rasterio
cligj==0.7.2
    # via
    #   fiona
    #   rasterio
defusedxml==0.7.1
    # via -r requirements.in
dotenv==0.9.9
    # via -r requirements.in
fiona==1.10.1
    # via -r requirements.in
flatbuffers==25.9.23
    # via tensorflow
gast==0.4.0
    # via tensorflow
geopandas==1.0.1
    # via
    #   -r requirements.in
    #   pygeoops
google-auth==2.41.1
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==1.0.0
    # via tensorboard
google-pasta==0.2.0
    # via tensorflow
grpcio==1.70.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.11.0
    # via tensorflow
idna==3.10
    # via requests
imageio==2.37.0
    # via scikit-image
importlib-metadata==8.5.0
    # via
    #   fiona
    #   markdown
    #   rasterio
keras==2.13.1
    # via tensorflow
lazy-loader==0.4
    # via scikit-image
libclang==18.1.1
    # via tensorflow
lxml==6.0.2
    # via -r requirements.in
markdown==3.7
    # via tensorboard
markupsafe==2.1.5
    # via werkzeug
networkx==3.2.1
    # via scikit-image
numpy==1.24.3
    # via
    #   -r requirements.in
    #   geopandas
    #   h5py
    #   imageio
    #   pandas
    #   pyogrio
    #   rasterio
    #   scikit-image
    #   scipy
    #   shapely
    #   snuggs
    #   tensorboard
    #   tensorflow
    #   tifffile
    #   topojson
oauthlib==3.3.1
    # via requests-oauthlib
opt-einsum==3.4.0
    # via tensorflow
packaging==25.0
    # via
    #   geopandas
    #   lazy-loader
    #   pyogrio
    #   scikit-image
    #   tensorflow
    #   topojson
pandas==2.0.3
    # via
    #   -r requirements.in
    #   geopandas
pillow==10.2.0
    # via
    #   -r requirements.in
    #   imageio
    #   scikit-image
protobuf==4.25.8
    # via
    #   tensorboard
    #   tensorflow
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pygeoops==0.4.0
    # via -r requirements.in
pyogrio==0.11.1
    # via geopandas
pyparsing==3.1.4
    # via snuggs
pyproj==3.5.0
    # via geopandas
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via dotenv
pytz==2025.2
    # via pandas
rasterio==1.3.11
    # via -r requirements.in
requests==2.32.4
    # via
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rsa==4.9.1
    # via google-auth
scikit-image==0.24.0
    # via -r requirements.in
scipy==1.10.1
    # via
    #   -r requirements.in
    #   scikit-image
shapely==2.0.7
    # via
    #   geopandas
    #   pygeoops
    #   topojson
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
snuggs==1.4.7
    # via rasterio
tensorboard==2.13.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.13.0
    # via -r requirements.in
tensorflow-estimator==2.13.0
    # via tensorflow
tensorflow-io-gcs-filesystem==0.34.0
    # via tensorflow
termcolor==2.4.0
    # via tensorflow
tifffile==2024.8.30
    # via scikit-image
topojson==1.10
    # via pygeoops
typing-extensions==4.5.0
    # via tensorflow
tzdata==2025.2
    # via pandas
urllib3==2.2.3
    # via requests
werkzeug==3.0.6
    # via tensorboard
wheel==0.45.1
    # via
    #   astunparse
    #   tensorboard
wrapt==1.17.3
    # via tensorflow
zipp==3.20.2
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
